Fixes: slowness on embedded devices.
Details: On embedded devices like n810 (which uses an Arm processor) usage
of double is incredably expensive.  In almost all cases our usage of double
can be done just as well using floats on those devices.
This is something that Qt does as well using the type 'qreal' which is a
double on all modern platforms but float on those that don't have double
natively.  So we get the best of both worlds by replacing double with qreal.

svn path=/trunk/koffice/; revision=851499