Added possibility of creating kexi reports elements in any direction
Added possibility of creating report elements no matter where mouse
was pressed and where it was released. Any direction will work. Also
minium size elements are moved to fit exactly the scene and do not go
out of it in case mouse release took place out of scene

BUG: 334967
FIXED-IN: 2.9
REVIEW: 120634