Add 'Value' property to report items which can contain static data if not bound to a field
Added a new property to the following report items:
check
field
text
barcode
Renamed image property 'Static Image' to Value but left saving/loading the same for compatibility.

Added a method to KoReportDesignerItemRectBase to store a string used for rendering data in the report designer.
Report designer shows the static data if used

Includes a bug fix in KRSize which meant recursion was possible
Includes a small fix to Text item which caused an infinite loop.

RVIEW: 120935