Use a non-legacy id also with id for action of the Line item
Summary:
Use a non-legacy id also with id for action of the Line item. `KoReportPluginManager::actions()` sets `act->setObjectName(metaData->id());`, so for consistency with the current ids it would be nice if also the id used for the Line item would have that namespace.

Reason is that Plan filters the actions to use by those ids, and for less confusion it would be great if the ids used would be consistent :)

Code is like that (in plan/libs/ui/reports/reportview.cpp, updated ids not yet committed):
```
    // allow only the following item types, there is not appropriate data for others
    QStringList itemtypes;
    itemtypes << "report:line"
        << "org.kde.kreport.field"
        << "org.kde.kreport.text"
        << "org.kde.kreport.check"
        << "org.kde.kreport.label"
        << "org.kde.kreport.chart"
        << "org.kde.kreport.web";
    QActionGroup *ag = new QActionGroup( this );
    int i = 0;
    foreach( QAction *a, m_designer->itemActions( ag ) ) {
        if ( itemtypes.contains( a->objectName() ) ) {
            QToolButton *tb = new QToolButton( w );
            tb->setObjectName( a->objectName() );
            tb->setIcon( a->icon() );
            tb->setText( a->text() );
```

Reviewers: piggz, staniek

Reviewed By: staniek

Differential Revision: https://phabricator.kde.org/D121