KReportPageFormat: align tr() extraction context and fetching context
Summary:
With Qt there are two different types of contexts for translated strings,
the "class" context and the "disambiguation" context, which are used together
to identify the string to pick from the catalog.
With QT_TRANSLATE_NOOP, the first argument is the "class" context and the
"disambiguation" context is 0. With QObject::tr(), the "class" context will
be automatically "QObject" and the second argument is the "disambiguation" context.
So
    QT_TRANSLATE_NOOP("Page size", "ISO A4")
    context -> "class": "Page size", "disambiguation": 0
and
    QObject::tr(pageFormatInfo[ format ].descriptiveName, "Page size")
    context -> "class": "QObject", "disambiguation": "Page size"
will not match.

So use "KReportPageFormat" as "class" context in all places and drop the
"disambiguation" context, as the "class" one should be hint enough to
translators and as to add the "disambiguation" one the macro
QT_TRANSLATE_NOOP3 seems to be needed, which returns a struct of two
const char*, so would bloat the pageFormatInfo data without real gain IMHO.

Reviewers: piggz, staniek

Reviewed By: staniek

Differential Revision: https://phabricator.kde.org/D135