Expand the documentation’s main page with an overview and example
Summary:
This may be something that only happens to me, but i find difficult to
use a new library when there is only documentation for each individual class
without a description of how they work together. In this case, between the
example and the fact that the library is not that big, it is relatively easy to
figure things out, but i still believe that it is harder than it could be.

I only modified MainPage.dox to show what are the main classes of KReport and
what are their jobs relative to each other within the library. They are a
cleaned up version of the notes i took while trying to integrate KReport to a
Qt application, so are written from the point of view of a first-time user and
only for these parts that i needed.

This is no excuse for not being knowing how to express things better; sorry for
that.

The code that i have written in the section for how to create plugins and data
sources is only in MainPage.dox because i thought that it would make the sample
application harder to follow if in the examples folder there are code from
multiple applications or libraries, but i certainly have created the projects
to test them all and i could add these files if it seems appropriate.

Reviewers: staniek, piggz

Reviewed By: staniek

Subscribers: staniek

Tags: #kreport

Differential Revision: https://phabricator.kde.org/D28809